Move code to load_balancer
